# Optimize structure of tracedoc