Now, a tracedoc has a `_lastversion` field which saves its all (old) values, and `_keys` and `_changes` field saves its changes(keys and values). The "newestversion" of a tracedoc SCATTERED in above 3 fields. As a result, tracedoc's `index` metamethod has to be a function, and `__len`, `__pairs`, `__ipairs` metamethods are hard to implement.
I try to optimize tracedoc for problems above.
- In commit `30f2b90`, a tracedoc's `index` metamethod becomes a table (its `_stage` field), which is its own "newestversion". On the other hand, its `_changed_keys` and `_changed_values` field saves its "lastversion" keys and values have been changed. Benefits from the `_stage` field, we can implement `__len`, `__pairs`, `__ipairs` metamethods much more simply.
- In commit `817c29f`, I export `tracedoc.next` to user, and add support for `tracedoc.concat`, with their test cases.