@jojo59516 2019-11-08
I've made a minimal benchmark between the current version and my optimized tracedoc:
- current version of tracedoc (commit `6ca7ed6`)
```git
❯ lua5.3 benchmark.lua
__index: time - 0.984 s
tracedoc.len: time - 0.891 s
tracedoc.ipairs: time - 1.906 s
tracedoc.pairs: time - 5.156 s

❯ luajit -joff benchmark.lua
__index: time - 0.859 s
tracedoc.len: time - 0.688 s
tracedoc.ipairs: time - 1.547 s
tracedoc.pairs: time - 3.375 s
```
- optimized version of tracedoc (commit `30f2b90`)
```git
❯ lua5.3 benchmark.lua
__index: time - 0.609 s
tracedoc.len: time - 0.703 s
tracedoc.ipairs: time - 1.016 s
tracedoc.pairs: time - 1.469 s

❯ luajit -joff benchmark.lua
__index: time - 0.719 s
tracedoc.len: time - 0.594 s
tracedoc.ipairs: time - 0.781 s
tracedoc.pairs: time - 1.078 s
```
with lua and luajit:
```git
❯ lua5.3 -v
Lua 5.3.3  Copyright (C) 1994-2016 Lua.org, PUC-Rio

❯ luajit -v
LuaJIT 2.1.0-beta3 -- Copyright (C) 2005-2017 Mike Pall. http://luajit.org/
```

Here is my benchmark code:
```lua
-- benchmark.lua
local function time(title, f)
    local clock = os.clock
    local time = 0
    local start = function() time = time - os.clock() end
    local stop = function() time = time + os.clock() end

    collectgarbage()
    f(start, stop)

    print(string.format("%s: time - %.3f s", title, time))
end

local tracedoc = require("tracedoc")
time("__index", function (start, stop)
    local doc = tracedoc.new {
        k = 'v', 
        1, 2, 3, 
    }

    local v
    for i = 1, 1000000 do
        start()
        v = doc.k
        v = doc[1]
        v = doc[0] -- nil
        v = doc[2] -- nil
        stop()
    end
end)

time("tracedoc.len", function (start, stop)
    local t = {}
    for i = 1, 10 do
        t[i] = i
    end

    local doc = tracedoc.new (t)

    local v
    for i = 1, 1000000 do
        start()
        v = tracedoc.len(doc)
        stop()
    end
end)

time("tracedoc.ipairs", function (start, stop)
    local t = {}
    for i = 1, 10 do
        t[i] = i
    end

    local doc = tracedoc.new (t)

    for i = 1, 1000000 do
        start()
        for k, v in tracedoc.ipairs(doc) do
            
        end
        stop()
    end
end)

time("tracedoc.pairs", function (start, stop)
    local t = {}
    for i = 1, 10 do
        t[i] = i -- array part
        t[-i] = -i -- hash part
    end

    local doc = tracedoc.new (t)

    for i = 1, 1000000 do
        start()
        for k, v in tracedoc.pairs(doc) do
            
        end
        stop()
    end
end)
```